Disable scrolling on body tag
If the user scrolls to the top of the results, then the scroll event
will bubble up to the body. The main app has position fixed though, so
it will appear to the user that scrolling is unresponsive for a few
moments.

The support for overscroll-behaviour is not widespread (safari) so
disabling touch events on the body is an attempt to skirt this.