Change install prompt
I had originally intended to implement the install prompt as a separate
route with react router however, react router doesn't play nicely with
github pages routing on the server side.

I have therefore implemented it as an overlay that will pop up
automatically for any navigation to the page that is not installed yet.