Use multiple db indexes
When retrieving results for a query, the order is important. By storing
exact matches in a separate index, we can do two different queries and
return the results as we see them. Since the exact matches will be first
we don't have to sort them and can serve them as-is.