Decompress dictionary before population
Since we are now serving the dictionaries in a gzipped compressed
format, we have to decompress them on the client.

Once they are decompressed, we can then populate the IndexedDB.

It makes sense to cache the compressed dictionary files, as we want to
make the app as offline enabled as possible.