Separate service worker from data load
Loading the data into the IndexedDB takes time, we want to let the user
decide when they will initiate this.

Previously we would initialte this in the serviceworker's install
method, because we could extend the event, but this is not a viable
solution. Now, we will cache the raw dictionary files in the sw, and
have the main thread load them into the IDB. Since the loading is done
asynchronously, it should not have a large impact on the user
experience.