Use gzip decompression stream
Using the decompression implementation that uses a buffer requires us to
keep the entire decompressed dictionary in memory as a blob. This will
exceed the memory quota on mobile safari.

By using the decompression stream, we can decompress and parse the JSON
as we download the bundle, keeping memory requirements low.